PR comments added. utils.py created
